reorganized file structure, initial setup is complete.
