+++ src/components/landing-div.jsx
@@ -1,0 +1,12 @@
+// import Navbar from "../navbar";
+
+
+function IntroDiv() {
+    return (
+        <div>
+            <h1>lading div entry point here</h1>
+        </div>
+    );
+  }
+  
+  export default IntroDiv;
