+++ src/pages/error.jsx
@@ -1,4 +1,4 @@
-import Navbar from "../navbar";
+import Navbar from "../components/navbar";
 
 
 function ErrorPage() {
